LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/pixel-manipulation/2d.imageData.put.unaffected-expected.txt

2d.imageData.put.unaffected

putImageData() is not affected by context state


PASS putImageData() is not affected by context state 
